requirements.txt
apache-airflow==2.8.1
mlflow==2.11.1
scikit-learn==1.4.0
pandas==2.1.4
numpy==1.26.3
psycopg2-binary==2.9.9
